program_aided_language_models/pal_requirements.txt
# langchain
langchain-experimental
openai==0.28
python-dotenv